+++ libft/srcs/ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: jgambard <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/15 13:00:02 by jgambard          #+#    #+#             */
-/*   Updated: 2020/02/15 13:00:30 by jgambard         ###   ########lyon.fr   */
+/*   Updated: 2020/03/09 07:05:34 by jgambard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
+#include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t		ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int		len;
-	unsigned int		i;
+	size_t		len;
+	size_t		i;
 
 	len = 0;
 	while (src[len])
